fix: address PR #37 review — move helpers to tests/helpers.py
- Move non-fixture helpers (mock_openai_response, create_test_job,
  cleanup_job, segment_factory) from conftest.py to tests/helpers.py
  to avoid pytest plugin module import issues
- Add status code assertions in create_test_job and cleanup_job
- conftest.py now only contains pytest fixtures (sample_segments,
  make_client)
